// src/types/info.tsx
export type infoProps = {
  heading: string;
  key1: string;
  profile: {
    firstName: string;
    lastName: string;
    bvn: string;
    gender: string;
  };
  key2: string;
  phoneNumber: string;
  key3: string;
  email: string;
  key4: string;
  education: {
    level: string;
    employmentStatus: string;
    duration: string;
    sector: string;
    officeEmail: string;
    monthlyIncome: {
      0: number;
      1: number;
    };
    loanRepayment: number;
  };
  key5: string;
  socials: {
    twitter: string;
    facebook: string;
    instagram: string;
  };
  key6: string;
  guarantor: {
    firstName: string;
    gender: string;
    phoneNumber: string;
    relationship: string;
  };
  key7: string;

  key8: string;

  key9: string;

  key10: string;
  value1: string;
  value2: string;
  value3: string;
  value4: string;
  value5: string;
  value6: string;
  value7: string;
  value8: string;
};